+++ dflib-csv/src/test/java/org/dflib/csv/CsvLoader_ImperfectSourceTest.java
@@ -6,8 +6,15 @@ import org.dflib.junit5.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class CsvLoader_ImperfectSourceTest {
+
+    static final ByteSource INCOMPLETE_ROWS = ByteSource.of("""
+            A,B,C
+            0,1,2
+            3,4
+            5""".getBytes());
 
     @Test
     public void duplicateColumnNames() {
@@ -29,4 +36,41 @@ public class CsvLoader_ImperfectSourceTest {
                 .expectRow(1, "4", "5", "6");
     }
 
+    @Test
+    public void incompleteRows() {
+        assertThrows(ArrayIndexOutOfBoundsException.class, () -> new CsvLoader().load(INCOMPLETE_ROWS));
+    }
+
+    @Test
+    public void incompleteRows_nullPadRows() {
+        DataFrame df = new CsvLoader()
+                .nullPadRows()
+                .load(INCOMPLETE_ROWS);
+
+        new DataFrameAsserts(df, "A", "B", "C")
+                .expectHeight(3)
+                .expectRow(0, "0", "1", "2")
+                .expectRow(1, "3", "4", null)
+                .expectRow(2, "5", null, null);
+    }
+
+    @Test
+    public void incompleteRows_nullPadRows_Primitives() {
+        assertThrows(IllegalArgumentException.class, () ->
+                new CsvLoader().nullPadRows().intCol("C").load(INCOMPLETE_ROWS));
+    }
+
+    @Test
+    public void incompleteRows_nullPadRows_PrimitiveDefaults() {
+        DataFrame df = new CsvLoader()
+                .intCol("C", -1)
+                .nullPadRows()
+                .load(INCOMPLETE_ROWS);
+
+        new DataFrameAsserts(df, "A", "B", "C")
+                .expectHeight(3)
+                .expectRow(0, "0", "1", 2)
+                .expectRow(1, "3", "4", -1)
+                .expectRow(2, "5", null, -1);
+    }
 }